+++ CHANGELOG.md
@@ -22,6 +22,9 @@ TombEngine releases are located in this repository (alongside with Tomb Editor):
 ### Lua API changes
 * Added missing constructor for `Collision.Probe` without room number.
 * Added optional looping argument for `View.GetFlybyPosition` and `View.GetFlybyRotation` functions.
+* Added EmitWaterfallMist under Effects.
+* Updated MakeExplosion and EmitBlood to spawn correct effects when used underwater.
+
 
 ## [Version 1.8](https://github.com/TombEngine/TombEditorReleases/releases/tag/v1.8) - 2025-03-16